refactor: Extract TerminalKeyboardMap and finalize
